refactor(ui): move reusable primitives into src/components
Per PROJECT_GUIDELINES.md, src/components/ holds reusable shadcn-style
UI components while features/ keeps feature slices:

- Move Modal and PasswordInput from features/shared to components/
- Extract the Toggle switch duplicated in SettingsTab and AdvancedTab
  into components/Toggle.tsx (keeping the type=button variant)
- features/shared keeps app-specific shared pieces only (Titlebar,
  StatusBar, ErrorToast, dialogs, announcement banner)
