+++ src/features/toolbox/AdvancedTab.tsx
@@ -4,6 +4,7 @@ import { useTranslation } from "../../lib/i18n";
 import { useConfigStore } from "../../lib/stores/config-store";
 import { useSetConfig } from "../../lib/hooks/use-config";
 import { commands } from "../../lib/tauri";
+import { Toggle } from "../../components/Toggle";
 
 export function AdvancedTab() {
   const { t } = useTranslation();
@@ -217,25 +218,5 @@ function Dropdown({
         </div>
       )}
     </div>
-  );
-}
-
-function Toggle({ checked, onChange }: { checked: boolean; onChange: () => void }) {
-  return (
-    <button
-      type="button"
-      role="switch"
-      aria-checked={checked}
-      onClick={onChange}
-      className={`relative h-5 w-9 shrink-0 rounded-full transition-colors ${
-        checked ? "bg-accent" : "bg-[var(--border)]"
-      }`}
-    >
-      <span
-        className={`absolute top-0.5 left-0.5 h-4 w-4 rounded-full bg-white transition-transform ${
-          checked ? "translate-x-4" : "translate-x-0"
-        }`}
-      />
-    </button>
   );
 }
