+++ src/features/toolbox/SettingsTab.tsx
@@ -4,6 +4,7 @@ import { useConfigStore } from "../../lib/stores/config-store";
 import { useSetConfig } from "../../lib/hooks/use-config";
 import { useUiStore } from "../../lib/stores/ui-store";
 import { commands } from "../../lib/tauri";
+import { Toggle } from "../../components/Toggle";
 import type { ThemeMode, Language } from "../../lib/stores/ui-store";
 
 const THEMES: { value: ThemeMode; labelKey: string }[] = [
@@ -170,24 +171,5 @@ function SettingRow({ label, children }: { label: string; children: React.ReactN
       <span className="text-xs font-semibold text-[var(--text)]">{label}</span>
       {children}
     </div>
-  );
-}
-
-function Toggle({ checked, onChange }: { checked: boolean; onChange: () => void }) {
-  return (
-    <button
-      onClick={onChange}
-      role="switch"
-      aria-checked={checked}
-      className={`relative h-5 w-9 shrink-0 rounded-full transition-colors ${
-        checked ? "bg-accent" : "bg-[var(--border)]"
-      }`}
-    >
-      <span
-        className={`absolute top-0.5 left-0.5 h-4 w-4 rounded-full bg-white transition-transform ${
-          checked ? "translate-x-4" : "translate-x-0"
-        }`}
-      />
-    </button>
   );
 }
